refactor(schemas): change reduce usage & move tags
