+++ src/common/types.ts
@@ -1,0 +1,1 @@
+export type Tags = Record<string, string | number | boolean | null>;
