+++ src/schema/providers/fileProvider/fileProvider.ts
@@ -6,8 +6,6 @@ import { ISchemaProvider } from '../provider';
 import { Schema } from '../../models/types';
 import { parseSchema } from '../../utils/schemaParser';
 
-export type Tags = Record<string, string | number | boolean | null>;
-
 @injectable()
 export class FileSchemaProvider implements ISchemaProvider {
   public constructor(@inject(SERVICES.CONFIG) private readonly config: IConfig) {}
